Fixed the locking problems of the context (moved it out of the mutex)

+++ src/pile.rs
@@ -1,7 +1,7 @@
 use alloc::{vec::Vec, boxed::Box};
 
 /// The address used by the pile to index logs (limits how many unique logs you can have)
-pub type Addr = u8;
+pub type Addr = u16;
 /// The type used to count the occurances of an address
 pub type AddrCounter = u8;
